Consolidated some functions from miniunz.c into unzip.c
